Clarify what intersect and difference return
The original descriptions said that the method `returns a new structure of the keys that...` which to me reads as different than what it returns (a `Collection`) since the `Collection` should have key/value pairs.  Left as-is, I wasn't sure if it returned an iterator or array or something else based just on the description.  This PR makes the description more explicit to remove ambiguity.