chore: clarify what intersect and difference return (#18)
* Clarify what intersect and difference return

The original descriptions said that the method `returns a new structure of the keys that...` which to me reads as different than what it returns (a `Collection`) since the `Collection` should have key/value pairs.  Left as-is, I wasn't sure if it returned an iterator or array or something else based just on the description.  This PR makes the description more explicit to remove ambiguity.

* Update descriptions to be more consistent

Rewrote `intersect` and `difference` descriptions to be clearer, and replaced other methods' and properties' descriptions to refer to `item(s)` where appropriate instead of several different terms.