feat: add 'ensure' method
Inspired by `[Enmap](https://enmap.evie.dev)#ensure`, `Collection#ensure` gets an element with the specified key if it exists, otherwise sets it to the provided `defaultValue` and returns the `defaultValue`. Useful for things like per-guild settings where you want to either get it or set it to a default value.